Refactor Shell ITs
* Create ShellCreateTableIT from tests moved from ShellServerIT
* Create TestShell class
* Modify ShellServerIT to not delete tables after ever test
* Move ShellConfigIT into shell package
* Move ShellIT into shell package
